test: mv unit test into polaris-server/services
