+++ service/README.md
@@ -1,1 +1,13 @@
-# Naming Service
+# 接口测试
+
+> 对业务逻辑进行完整性测试
+
+## 依赖
+
+* 数据库地址，配置一个测试的数据库地址
+
+## 执行
+
+```
+go test -v
+```
