+++ service/l5service_test.go
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package discover
+package service
 
 // . "github.com/smartystreets/goconvey/convey"
